Putting the users answers into the constructor elements.
Used require() to import the other js pages. All the promise functions to also add a new key into each element of the array - role. Then I used map to go through each of the elements in the array and if they equaled to Manager - then use the constructor to create a manager element. And so on for each of the intern and engineers.