updated and copied the HTML I will be using into the index.
Updated the HTML (removed the icons because it was breaking the html? might troubleshoot and add back in later). Copied the HTML I will be using into the index.js -> and seperated out the Manager, Enginner, and Inter HTML into variables (still need to create variables and call in html to generate the users answers.)